move userview to add userfriends
