+++ frontend/src/components/userview/UserFriends.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function UserFriends() {
+  return (
+    <div>UserFriends</div>
+  )
+}
+
+export default UserFriends
